refactor: Core - View 분리
