+++ Projects/MainFeature/Sources/Views/Components/CommentList/CommentListView.swift
@@ -1,26 +1,6 @@
 import SwiftUI
-import ComposableArchitecture
 
-public struct CommentListCore: Reducer {
-  public struct State: Equatable {
-    let voteID: Int
-  }
-  
-  public enum Action: Equatable {
-    
-  }
-  
-  @Dependency(\.network) var network
-  
-  public var body: some ReducerOf<Self> {
-    Reduce { state, action in
-      switch action {
-      default:
-        return .none
-      }
-    }
-  }
-}
+import ComposableArchitecture
 
 public struct CommentListView: View {
   let store: StoreOf<CommentListCore>
